Add files via upload
In the era of social media and networking, the usage of bad words and aggressive words has increased significantly. The young population is playing a major role in it. Cyberbullying affects more than half of the young population using social media. Insults on social media websites create negative interactions within the network. These remarks build up a culture of disrespect in cyberspace. Algorithms and tools used to understand and mitigate it are mostly inactive. Also, current implementations of insult detection using machine learning and natural language processing have very low recall rates. In short, the paper involves determining ways to identify bullying in the text by analyzing and experimenting with different methods to find a feasible way of classifying such comments. We proposed an efficient algorithm to identify the bullying test and aggressive comments and analysed these comments to check their validity. NLP and Machine learning is used for analyzing social comment and identifying the aggressive effect of an individual or a group. An effective classifier acts as the core component in a final prototype system that can detect cyberbullying on social media.